visualization: Remove channel and only render on MainEventsCleared
With the loop and channel we were previously redrawing and rendering a
frame _for every event_.  With vsync enabled this immediately locks up
the application as mouse polling rates are well above that.

We cleaned our own applications to embrace winit's "new" eventloop
closure instead of dealing with all the desync, race conditions and
other oddities when using a separate thread to handle the events, and do
the same here too.